+++ src/Components/Footer.js
@@ -1,3 +1,15 @@
-export default function Footer() {
-  return;
+import React from "react";
+
+function Footer() {
+  return (
+    <footer>
+      <ul>
+        <li>Contact</li>
+        <li>About</li>
+        <li>Legal</li>
+      </ul>
+    </footer>
+  );
 }
+
+export default Footer;
