+++ src/Components/Header.js
@@ -1,3 +1,11 @@
-export default function Header() {
-  return;
+import React from "react";
+
+function Header() {
+  return (
+    <header>
+      <h1 className="color-change-5x">RSVP App</h1>
+    </header>
+  );
 }
+
+export default Header;
